# [coverage] Speed up all tests x2