--- a/test/configs/travis.utsc
+++ b/test/configs/travis.utsc
@@ -1,0 +1,11 @@
+{
+  "testfiles": [
+    "*.uts",
+    "../scapy/contrib/*.uts"
+  ],
+  "onlyfailed": true,
+  "preexec": {
+    "../scapy/contrib/*.uts": "load_contrib(\"%name%\")"
+  },
+  "format": "text"
+}