--- a/test/configs/windows2.utsc
+++ b/test/configs/windows2.utsc
@@ -1,0 +1,15 @@
+{
+  "testfiles": [
+    "*.uts",
+    "..\\scapy\\contrib\\*.uts"
+  ],
+  "onlyfailed": true,
+  "preexec": {
+    "..\\scapy\\contrib\\*.uts": "load_contrib(\"%name%\")"
+  },
+  "format": "html",
+  "kw_ko": [
+    "combined_modes_ccm",
+    "mock_read_routes6_bsd"
+  ]
+}