This PR:
- brings a new feature: multi file support in UTscapy.
That allow us to use the same scapy instance to run multiple tests, and speed up the tests. (also added a new log level with more user-friendly texts)
- brings a new feature: JSON config files in UTscapy. Optional, it reduces the command line complexity and allow to use complex `PREEXEC` structures
- Fixes some bugs with the tests (`_` used with `xrange` made `_` a static value, removed `sys.__stdout__` usage)
- Fixes `fixname()` function not being called in DaDict

(PR won't work without the bug fixes)

Results:
- Travis CI:
from an [older build](https://travis-ci.org/gpotter2/scapy/builds/207658839), built in 9 min 19 sec, we're [now](https://travis-ci.org/gpotter2/scapy/builds/207800650) at 4 min 48 sec
- Appveyor: even better,
from [older](https://ci.appveyor.com/project/gpotter2/scapy/build/1.0.296) in 9 min 21 sec we're [now](https://ci.appveyor.com/project/gpotter2/scapy/build/1.0.298) at 3 min 56 sec.

So in average, that should speed up the tests (running part only), by 2.