@p-l- 2017-03-10
Nice @gpotter2. Just curious: what was the problem with renaming `.update()`?