@gpotter2 2017-03-10
I thought maybe simply by renaming it `.update()`, calling `self.__dict__.update` would have used it... That wasn't the case